src/requirements.txt
flask
nltk==3.3
python-twitter==3.5
